auto/requirements.txt
langgraph==0.2.59
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.7
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2
duckduckgo-search==6.3.7